+++ src/main/java/de/markusbordihn/cats/commands/CatCommands.java
@@ -40,6 +40,7 @@ public final class CatCommands extends AbstractCommandCollection {
     register(new CatInfoCommand());
     register(new CatListCommand());
     register(new CatNameCommand());
+    register(new CatNeedsCommand());
     register(new CatOwnerCommand());
     register(new CatPlayCommand());
     register(new CatReloadCommand());