+++ src/main/java/de/markusbordihn/cats/handler/NPCSetupHandler.java
@@ -26,6 +26,7 @@ import com.hypixel.hytale.server.npc.asset.builder.BuilderFactory;
 import com.hypixel.hytale.server.npc.instructions.Action;
 import com.hypixel.hytale.server.npc.instructions.Sensor;
 import de.markusbordihn.cats.actions.BuilderActionCatFetchYarnBall;
+import de.markusbordihn.cats.actions.BuilderActionCatGoToBed;
 import de.markusbordihn.cats.actions.BuilderActionCatIdleChoice;
 import de.markusbordihn.cats.actions.BuilderActionCatInteractionBase;
 import de.markusbordihn.cats.actions.BuilderActionCatInteractionOwner;
@@ -147,6 +148,14 @@ public class NPCSetupHandler {
     } catch (Exception e) {
       LOGGER.at(Level.SEVERE).log(
           "Failed to register action: %s", BuilderActionCatTeleportToBed.BUILDER_ID, e);
+    }
+
+    try {
+      actionFactory.add(BuilderActionCatGoToBed.BUILDER_ID, BuilderActionCatGoToBed::new);
+      LOGGER.at(Level.INFO).log("Registered action: %s", BuilderActionCatGoToBed.BUILDER_ID);
+    } catch (Exception e) {
+      LOGGER.at(Level.SEVERE).log(
+          "Failed to register action: %s", BuilderActionCatGoToBed.BUILDER_ID, e);
     }
 
     try {